test/fix(harnesses): address Polly review — clearer non-2xx log, shared test helper, unknown-event guard
Non-blocking follow-ups from the Polly AI review on PR #579:

- Log non-2xx responses distinctly from connection errors. Both native
  hooks now catch httpx.HTTPStatusError before the broad httpx.HTTPError
  branch and log the status code, so a real AP outage (e.g. 503) is
  distinguishable from an unreachable server in production diagnostics.
  Behavior is unchanged — both still fail closed for the tool-call gate.
- Deduplicate the failing-client test stub into
  tests/native_hook_helpers.make_failing_client, imported by both the
  claude- and codex-native hook test modules, so the four failure modes
  can't drift.
- Add an explicit unknown-event test for fail_closed_hook_output
  ("SomeNewEvent" -> None) documenting the fail-open-for-unknowns contract.
