+++ tests/test_claude_native_hook.py
@@ -19,6 +19,7 @@ from omnigent.claude_native_bridge import (
     record_hook_event,
     write_active_session_id,
 )
+from tests.native_hook_helpers import make_failing_client
 
 
 @pytest.fixture(autouse=True)
@@ -1711,40 +1712,6 @@ def test_ask_user_question_hook_returns_deny_without_updated_input(
     )
 
 
-def _make_failing_client(mode: str) -> type:
-    """
-    Build an httpx.Client stub that fails the policy POST a given way.
-
-    :param mode: One of ``"connect_error"`` (POST raises), ``"non_2xx"``
-        (503 → ``raise_for_status``), ``"empty_body"`` (200, no content),
-        or ``"malformed_json"`` (200, non-JSON body).
-    :returns: A class usable as a drop-in for :class:`httpx.Client`.
-    """
-
-    class _FailingHttpxClient:
-        def __init__(self, *, headers: dict[str, str], timeout: object) -> None:
-            del headers, timeout
-
-        def __enter__(self) -> _FailingHttpxClient:
-            return self
-
-        def __exit__(self, *args: object) -> None:
-            del args
-
-        def post(self, url: str, *, json: dict[str, object]) -> httpx.Response:
-            del json
-            req = httpx.Request("POST", url)
-            if mode == "connect_error":
-                raise httpx.ConnectError("AP unreachable", request=req)
-            if mode == "non_2xx":
-                return httpx.Response(503, text="upstream down", request=req)
-            if mode == "empty_body":
-                return httpx.Response(200, content=b"", request=req)
-            return httpx.Response(200, text="not json at all", request=req)
-
-    return _FailingHttpxClient
-
-
 @pytest.mark.parametrize("mode", ["connect_error", "non_2xx", "empty_body", "malformed_json"])
 def test_evaluate_policy_pre_tool_use_fails_closed_when_verdict_unavailable(
     tmp_path: Path,
@@ -1761,7 +1728,7 @@ def test_evaluate_policy_pre_tool_use_fails_closed_when_verdict_unavailable(
     """
     monkeypatch.setattr("omnigent.claude_native_bridge._TRUSTED_PARENT", tmp_path)
     monkeypatch.setattr("omnigent.claude_native_bridge._BRIDGE_ROOT", tmp_path / "root")
-    monkeypatch.setattr(claude_native_hook.httpx, "Client", _make_failing_client(mode))
+    monkeypatch.setattr(claude_native_hook.httpx, "Client", make_failing_client(mode))
     bridge_dir = prepare_bridge_dir("conv_abc", bridge_id="bridge_shared", workspace=tmp_path)
     write_active_session_id(bridge_dir, "conv_active")
     build_hook_settings(bridge_dir, ap_server_url="http://127.0.0.1:8787")
@@ -1808,7 +1775,7 @@ def test_evaluate_policy_non_tool_call_phases_fail_open_on_error(
     """
     monkeypatch.setattr("omnigent.claude_native_bridge._TRUSTED_PARENT", tmp_path)
     monkeypatch.setattr("omnigent.claude_native_bridge._BRIDGE_ROOT", tmp_path / "root")
-    monkeypatch.setattr(claude_native_hook.httpx, "Client", _make_failing_client("connect_error"))
+    monkeypatch.setattr(claude_native_hook.httpx, "Client", make_failing_client("connect_error"))
     bridge_dir = prepare_bridge_dir("conv_abc", bridge_id="bridge_shared", workspace=tmp_path)
     write_active_session_id(bridge_dir, "conv_active")
     build_hook_settings(bridge_dir, ap_server_url="http://127.0.0.1:8787")
