+++ tests/test_codex_native_hook.py
@@ -18,6 +18,7 @@ from omnigent.codex_native_bridge import (
     write_bridge_state,
     write_policy_hook_config,
 )
+from tests.native_hook_helpers import make_failing_client
 
 
 class _DenyHttpxClient:
@@ -392,40 +393,6 @@ def test_missing_policy_config_is_fail_open(
     assert captured.out == ""
 
 
-def _make_failing_client(mode: str) -> type:
-    """
-    Build an httpx.Client stub that fails the policy POST a given way.
-
-    :param mode: One of ``"connect_error"`` (POST raises), ``"non_2xx"``
-        (503 → ``raise_for_status``), ``"empty_body"`` (200, no content),
-        or ``"malformed_json"`` (200, non-JSON body).
-    :returns: A class usable as a drop-in for :class:`httpx.Client`.
-    """
-
-    class _FailingHttpxClient:
-        def __init__(self, *, headers: dict[str, str], timeout: object) -> None:
-            del headers, timeout
-
-        def __enter__(self) -> _FailingHttpxClient:
-            return self
-
-        def __exit__(self, *args: object) -> None:
-            del args
-
-        def post(self, url: str, *, json: dict[str, object]) -> httpx.Response:
-            del json
-            req = httpx.Request("POST", url)
-            if mode == "connect_error":
-                raise httpx.ConnectError("AP unreachable", request=req)
-            if mode == "non_2xx":
-                return httpx.Response(503, text="upstream down", request=req)
-            if mode == "empty_body":
-                return httpx.Response(200, content=b"", request=req)
-            return httpx.Response(200, text="not json at all", request=req)
-
-    return _FailingHttpxClient
-
-
 @pytest.mark.parametrize("mode", ["connect_error", "non_2xx", "empty_body", "malformed_json"])
 def test_pre_tool_use_fails_closed_when_verdict_unavailable(
     bridge_dir: Path,
@@ -441,7 +408,7 @@ def test_pre_tool_use_fails_closed_when_verdict_unavailable(
     CLOSED (deny) instead of "no opinion" — the bypass reported in #536.
     """
     write_policy_hook_config(bridge_dir, ap_server_url="http://127.0.0.1:8787", ap_auth_headers={})
-    monkeypatch.setattr(codex_native_hook.httpx, "Client", _make_failing_client(mode))
+    monkeypatch.setattr(codex_native_hook.httpx, "Client", make_failing_client(mode))
 
     exit_code = _run_hook(
         bridge_dir,
@@ -486,7 +453,7 @@ def test_non_tool_call_phases_fail_open_on_error(
     runner-side ``FAIL_CLOSED_PHASES`` (PR #163).
     """
     write_policy_hook_config(bridge_dir, ap_server_url="http://127.0.0.1:8787", ap_auth_headers={})
-    monkeypatch.setattr(codex_native_hook.httpx, "Client", _make_failing_client("connect_error"))
+    monkeypatch.setattr(codex_native_hook.httpx, "Client", make_failing_client("connect_error"))
 
     exit_code = _run_hook(bridge_dir, payload, monkeypatch)
